requirements.txt
django==2.2
django_chartit